feat: Add parameters, lambda and refactor
- Add parameters for slack webhook (in aggregator account)
- Add lambda in aggregator account
- Refactor: Move stacks/constructs to separate files
