+++ src/Parameters.ts
@@ -1,0 +1,21 @@
+import { Construct } from 'constructs';
+import { StringParameter } from 'aws-cdk-lib/aws-ssm';
+
+
+export class Parameters extends Construct {
+  constructor(scope: Construct, id: string) {
+    super(scope, id);
+
+    const slackParam = new StringParameter(this, 'ssm_slack_1', {
+      stringValue: '-',
+      parameterName: Statics.ssmSlackWebhookUrl,
+    });
+    slackParam.applyRemovalPolicy(RemovalPolicy.DESTROY);
+
+    const slackParamLowPriority = new StringParameter(this, 'ssm_slack_2', {
+      stringValue: '-',
+      parameterName: Statics.ssmSlackWebhookUrlLowPriority,
+    });
+    slackParamLowPriority.applyRemovalPolicy(RemovalPolicy.DESTROY);
+  }
+}
